LP bug#683507: xtrabackup should be updated to use the latest 5.1, 5.5                and XtraDB releases
Updated xtrabackup's Makefile and the 5.1/plugin patches so that they
are now based on MySQL 5.1.52. 

In addition, fix_innodb_for_backup_5.1_plugin_1.0.7.patch was renamed to
fix_innodb_for_backup_5.1_plugin.patch, since the plugin version is
different now, and it does not make sense to keep the version number in
the patch name.

This patch also fixes LP bug#681789, since I had to update and test the
broken patch anyway.