InnoDB compression support for incremental backups.
xtrabackup did not support compressed tables when doing incremental
backups. The reason for that was missing information about the page size
in incremental .delta files, so it was always assumed to be the default
one (16 KB) at the --prepare stage.

This patch removes that limitation by introducing per-delta meta
information which is stored in a file with the ".meta" extension and the
same name as the corresponding .delta file. Currently a .meta file can
only contain the page size, but we may add more fields to it later.