Bug #606981: linux + o_direct + stream backup = broken?
Streaming InnoDB files with tar4ibd could lead to filesystem hangs when
InnoDB was configured to access data files with the O_DIRECT flag. The 
reason was that tar4ibd did not have support for O_DIRECT and
simultaneous O_DIRECT + non-O_DIRECT access to a file on Linux is 
disallowed.

Fixed innobackupex and tar4ibd to use O_DIRECT on input InnoDB files if 
the value of innodb_flush_method is O_DIRECT in the InnoDB configuration.