Fix info.page_size may be used uninited warning.
xtrabackup.c: In function xtrabackup_copy_datafile:
xtrabackup.c:2451: warning: info.page_size may be used uninitialized in this function [-Wuninitialized]

I audited the code path and it is actually safe, but this is only because
of info only being used in incremental backups.
By initing to zero we avoid the warning.