Bug #798488: xtrabackup ignores defaults-file in apply-log-only prepare mode
For a number of reasons load_defaults() must be called before parsing
command line arguments. Fixing that MySQL limitation within xtrabackup
patches does not look reasonable.

The fix is to make xtrabackup fail with an error if --defaults-file is
not the first option in the command line (and thus, would be
ignored).

There are also some more minor cleanups in the test suite.