Bug #766607: new InnoDB table is not included to full prepared backup              from incremental
The problem was that when a new table was added after taking a full
backup *and* the corresponding pages had been flushed from the buffer
pool before taking an incremental one, the new table data went to a
.delta file. When trying to apply that .delta to a full backup later,
xtrabackup_apply_delta() failed because the corresponding data file
could be found.

Fixed by creating data file (and directory, if necessary) if it is not
found when applying delta.

There is also a usual batch of test suite cleanups.