Bug #811065: xb_log_overwrap test hangs sporadically
The problem was introduced with a combination of the fix for bug #805593
and recent debug build fixes.

Since log_sys->mutex is now acquired to fix debug builds, it should also
be released when failing with an error.

Fixed by making sure the mutex is always released on error conditions.