Bug #722638: xtrabackup: taking backup while tables are droped and              created breaks backup
The root cause of the problem was that xtrabackup iterated .ibd files
based on a data dictionary snapshot taken at the very start of a backup
process. If a tablespace id changed due to dropping and recreating the
table with the same name, or an ALTER TABLE that requires a table copy,
or TRUNCATE (if it was performed as DROP + CREATE TABLE), InnoDB
complained about id mismatch, and crashed with an assertion failure.

The idea of the fix is to ignore tablespaces with wrong ids at the
backup stage. They will be (re)created when doing recovery at the prepare
stage, since the corresponding delete/create log records will be
replayed on recovery.

Fixed by modifying fil_node_open_file() in all XtraBackup patches to
InnoDB/XtraDB source to signal id mismatches back to the caller (with an
descriptive warning) rather than crash with an error. 

Warning messages in xtrabackup.c were also adjusted to take this new
case into account.

This patch also introduces a new command line switch to xtrabackup,
--debug-sync, to make possible automatic testing of the fix.