Fixed bug 857788
In some cases certain combinations of options would create a situation where innobackupex attempted to send commands through a mysql connection that was no longer open.

This patch greatly simplifies the connection open and closing so that it's less confusing to know, depending on which options you are using, whether you need to open a fresh connection or not, just always try to reuse the same one.