Bug #884737: xtrabackup's --parallel option asserts / crashes with a              value of -1
The problem was that my_getopt was implicitly converting negative
values to large positive ones, so xtrabackup would fail with an
asserts / crashes due to resource shortage.

Fixed by changing 'parallel' to a signed option so negative values are
detected and adjusted to the default value of 1 by my_getopt.