Bug 891496: tar4ibd crashes on multi-file tablespace files
The problem was that tar4ibd expected to read page size from the FSP
header of each data file, which, in case of a multi-file system
tablespace, is only available in the first file, but not in subsequent
ones. That resulted in tar4ibd using a bogus page size, hence the crash.

Fixed by enforcing UNIV_PAGE_SIZE as the page size for system tablespace
files.