Bug #917823: xtrabackup: error: log block numbers mismatch
The problem was then when the log copying thread processed the log
past its tail, it would fail with an error in case of an incompletely
written log block. Which is a regression introduced in the 2.0 branch by
the fix for bug #805593.

Fixed xtrabackup_copy_logfile() so that log copying only fails if
the read LSN is greater than expected, which is the logic used in 1.6.

No test case, as there is no way to test this automatically (would
require some kind of synchronization between the server, Xtrabackup and
the test itself).