Fix for bug #924026: innobackupex fails during backup if table is                      removed while copying
The problem was that when copying non-InnoDB files in innobackupex, a
table can conceivably exist at the point in time where the list of files
to copy is pulled, but have been removed before the attempt to copy the
file is made. For example, a table that exists for a short amount of
time.

Fixed by changing innobackupex gracefully handle this, by only throwing
an error if the copy failed and the file actually exists.