Bug #900175: Operating system error number 17 when use parallel backup
The problem was a race condition in datafiles_iter_next() resulting in
symptoms described in the bug report and sporadic test failures in
Jenkins: even though it->node modifications were protected by a mutex in
datafiles_iter_next(), it was also used as a return value (which
obviously occurred after released the mutex). Since the iterator context
may be shared between many threads with --parallel > 1, it->node might
be overwritten after the mutex has been released and return from the
function. So we might end up with multiple thread receiving the same
node from datafiles_iter_next().

Fixed by using a local variable as a return value instead of a shared
structure member.