Bug #977995: xbstream sometimes fails while extracting
The problem was in a partial read in xb_stream_read_chunk() when reading
the constant part of the chunk header. Unlike other parts, that one was
being read without the MY_FULL_IO flag to my_read(). So it could lead to
a partial read which wasn't handled correctly.

Fixed by looping over my_read() to read the necessary amount of bytes,
or bail out with an error in case of EOF.

Another problem was in a wrong constant being used to check if the read
was partial or not. Fixed by a cleanup to make sure the same constant is
always used, and some debug assertions.