Bug #977998: --stream=tar fails (archive_write_header failed)
The problem was in the USTAR format used by XtraBackup which is
incompatible with files larger than 8 GB.

Fixed by changing the libarchive format to 'restricted PAX' which avoids
the limitation. However, the PAX format generates some annoying (and
benign) warnings when the GNU tar format is used due to extended
attributes written to the tar stream by libarchive in the PAX mode. This
problem was fixed by modifying the bundled libarchive code to suppress
the extended attributes.