Bug #1002688: innobackupex incremental apply-log copies to wrong directory
The problem was that the copy_dir_recursively() routine relied on both
the source path and the target one to be in the canonical format,
i.e. not having any trailing slashes. However, when copying files after
merging an incremental backup to the full backup directory, the target
path (i.e. the full backup directory) was always in the canonical
format, because innobackupex calls File::Spec->rel2abs() on it in
check_args(), but the source path (i.e. the --incremental-directory
value) might not be in the canonical format in case the user specified
it with a trailing slash.

Fixed by canonizing both source and the target directories in
copy_dir_recursively() so it does not rely on the arguments format.