Bug #977101: --safe-slave-backup results in incorrect binlog info
The fundamental problem is in the way innobackupex talks to the
server. All queries are sent using the mysql command line client, then
the response is stored to file that is common for all queries. Then it
parses that file to figure out what might look as a response to the
specific query that was last issued. In this specific case, when
--safe-slave-backup was present, innobackupex confused the response from
SHOW SLAVE STATUS with the one from SHOW MASTER STATUS.

Since refactoring or deprecating innobackupex is not on our short-term
plans, fixed the bug by changing the SHOW MASTER STATUS output to the
vertical format and improving the parsing code so that we at least
capture the SHOW MASTER STATUS output rather than something else.

Added a test case using the new replication support in the test suite.