Bug #977652: Compressed/Uncompressed incremental backup fails on              compressed full backup in Xtrabackup 2.0.0
The problem was that xtrabackup created a compressed
xtrabackup_checkpoints file when compresses was used to create the full
backup. Which made full backups unusable for incremental backups without
decompressing them first.

Fixed by omitting compression for xtrabackup_checkpoints. 

A bit hacky implementation, a more clean fix will be possible in 2.1
with refactoring implemented in that branch.