Bug #996483: False positive testsuite failures with grep 2.10
The problem was in spurious test failures on Ubuntu 12.04 due to the
mentioned grep version being buggy: grepping a file and redirecting the
output to the same file would fail with an error even when the -q switch
is used, i.e. no output is actually generated.

Fixed by implementing a workaround with the grep output always
redirected to a pipe and then cat redirecting it back to standard
output, i.e. the original file, if the standard output was redirected.