Bug #1015416: Regression in XtraBackup 2.0.0 in case of an incomplete log block
The problem was that when XtraBackup was reading a log block that has not
been completely written by the server yet, it would fail with an error
complaining about log block checksum mismatch. Which is a regression
introduced in 2.0 by the fix for bug #805593.

Fixed by restoring the 1.6 behavior, i.e. just print a warning and retry
reading the log block later.