Bug #803636: "moves files" option needed with --copy-back
Merged the patch contributed by Igor Tverdovskiy in
lp:~igor-tverdovskiy/percona-xtrabackup/move-back_opt with some edits.

The patch adds a new option --move-back to innobackupex, which is
similar to --copy-back with the only difference that instead of copying
files it moves them to the target locations. It may be useful in cases
when there is not enough free disk space to hold both data files and
their backup copies.