Bug #1037379: SQL_THREAD left in stopped state of               safe-slave-backup-timeout is reached
Bug #887803: innobackupex --safe-slave-backup-timeout option doesn't
             work

The problem in bug #1037379 was that wait_for_safe_slave() in
innobackupex ran START/STOP SLAVE SQL_THREAD in a loop. If it succeeded
in getting slave_open_temp_tables to 0 in a specified time out, the
SQL thread was then started later after copying non-transactional
data. However, when it reached the timeout it did not restart the SQL
thread before terminating with an error.

Another problem was that if wait_for_safe_slave() succeeded,
innobackupex would start the SQL thread regardless of its initial
state. Which might be an undesired side effect.

Fixed by:

1. Checking the initial SQL thread state and starting it before
terminating with a timeout error in wait_for_safe_slave().

2. Starting the SQL thread only if it was running initially in backup()

Additionally, this revision also fixes bug #887803, because the test
case for bug #1037379 requires a working --safe-slave-backup-timeout
option.