Bug #1065561: xtrabackup fails on apply-log when filesystem doesn't               support linux aio Bug #1068459: xtrabackup ignores innodb_use_native_aio
The problem was that --prepare used a default value for
srv_use_native_aio (TRUE). Though some filesystems (e.g. tmpfs or
zfsonlinux) do not support it. Additionally the workaround of using an
explicit innodb_use_native_aio didn't work, because xtrabackup did not
recognize that option and silently ignored it.

Fixed both problems by adding the corresponding option and making it
disabled by default. Ideally the fix would be to test both datadir and
tmpdir for AIO support and disabling it if they don't on --prepare or
--stat. However, simply disabling AIO on --prepare would be sufficient
in 99% of cases. And it can now be enabled enabled back for the remaining
1%.