Introduce parallel test runner for Percona XtraBackup.
This uses a (relatively) simple C program to run several of the shell
test runners in parallel, serializing the subunit output so it can still
be consumed by other utilities. We use simple UNIX process control to control
everything, completely avoiding the horror of threads.
      
The xb_incremental_compressed test is split up from a big for loop as then
it can be run in parallel.