Switch to the debian source format "3.0 (native)"
In the debian source, switch from the 3.0 (quilt) to 3.0 (native) format.

The 3.0 (native) format is more suited for our packaging system. This is
an extension from the 1.0 format we had been using before upgrading and
it's good for cases when the debian/ buliding files are actually part of
the distribution, i.e., the debian/ maintainers are the same than the
maintainers of the package, The 3.0 (quilt) format is more suited for
cases where there's an upstream package maintained separately from the
debian packaging. Clearly, the former is the case for Percona XtraBackup.

Please refer to the dpkg-source(1) documentation for more information
about the differences in packaging.