Add an option in the debian builder to not build transitional packages.
Add a --notransitional option in the debian building script in order not
to build the xtrabackup transitional package.

The xtrabackup package was intended for having an easy transition when
the rebranding to percona-xtrabackup happened at revno 403, but it won't
be needed anymore in distros where the package is new and wasn't ever
called xtrabackup.