Bug #924492: Sporadic "buf_LRU_old_ratio >= 51" assertion failures in              5.1 debug builds in Jenkins
The problem was that InnoDB relies in buf_LRU_old_ratio to be always
initialized, and initializes it in ha_innodb.cc. However, XtraBackup
does not use code in ha_innodb.cc, so it did not initialize that
variable. 

As a result, the first call to buf_LRU_old_adjust_len() triggered a
debug-only assertion. That function is not normally called by XtraBackup
in a test suite run (it requires a certain level of pages in the buffer
pool).

InnoDB in 5.5+ also initializes that variable to a hard-coded constant
(and then adjusting it based on the actual server variable
value). That's why the problem only occurred in 5.1 debug Jenkins
builds.

The fix is to do the same for 5.1 builds: initialize the variable to a
hard-coded constant before starting an InnoDB instance in XtraBackup.