Fix bug 1098498 (Sporadic xb_incremental_compressed_* failures in Jenkins) and bug 1132763 (No need to shutdown servers cleanly before wiping their datadirs in the testsuite).
Most tests shutdown the server cleanly before copying back the
restored datadir.  These shutdowns:
1) may timeout sporadically (bug 1098498);
2) are not required in the first place (bug 1132763), because the
server data is about to be wiped, thus kill -9 would work just as
fine.

Fixed by replacing stop_server() implementation to do kill -9 instead
of a clean shutdown.  Add a new function shutdown_server() that is
similar to the old stop_server() implementation.  Adjust the testcases:
1) ib_part_*: wipe the server dir, restore not only test database, but
all of the databases.  For the testcases that work by selecting a list
of tables to backup, make sure to include all of mysql and
performance_schema tables.
2) bug766033: replace the database corruption of poking one zero byte
to the first page with overwriting the start of the second page with
0xAAAAAAAA to reduce the risk of the previous corruption attempt not
corrupting anything due to the zero byte already existing at the
location.
3) bug766607: rewrite the testcase to use innobackupex instead of
xtrabackup so that it works with the wiped server data dir.
4) xb_incremental: remove the testcase as it becomes identical to
ib_incremental atfer the changes to support wiped server data dir.
5) xb_partial: likewise, becomes identical to ib_part_*.
6) xb_export: use shutdown_server instead of kill_server for syncing
with the end of a tablespace import.

Also remove an unused function stop_all_servers() in common.sh and fix
a typo in testrun.sh.