Bug #1130627: Can't backup individual partitions
The bug is a regression introduced in 2.0.5 by the fix for bug
#711166. The fix introduced partition suffix ('#P#...') trimming in
tables names before checking them against regular expressions specified
by the --include option in innobackupex or the --tables option in
xtrabackup. Which made impossible partial backups of partitioned tables.

Fixed in both innobackupex and xtrabackup by first checking the table
name without the suffix trimmed, and then repeating the check with the
suffix trimmed if the former check fails. 

This patch also fixes bug #1131084 "Unneccessary/debug print in
xtrabackup output" and a few other minor issues introduced by the fix
for bug #711166.