Fix the binary builder so it exits if a build fails
The binary builder was failing to exit if a build failed, due to the way
bash interprets how set -e should be acted upon when a command fails. If
one of the build.sh commands failed, the others wouldn't execute, however
the tar file would be constructed anyway and the script would exit with
a successful status.

Fix this so when a build fails the directories would be erased, the tar
won't be created and the script will exit with a nonzero status.