--dummy for the dpkg builder moved to debian/rules
The --dummy option for the dpkg builder used a patch file to modify
debian/rules so it would build the dummy binaries. This proved cumbersome to
maintain since if the rules files changed then the patch would have to be
regenerated complicating commits.

Change it so debian/rules is aware of the DEB_DUMMY variable and acts
accordingly. The DEB_DUMMY variable is set from the dpkg builder script. This
is how the RPM dummy parameter works.