Update dpkg to current standards
Update Standards-Version to 3.9.3. For debian squeeze, this means that lintian
from squeeze-backports should be used, but since wheezy it's right around the
corner this is appropriate.

Update compat level to 7.

Use the dh scripts for building in debian/rules, with override targets. Remove
the obsolete debian/info file.

Move the /usr/share/percona-xtrabackup-test to the percona-xtrabackup-test
package.