Bug #1170806: innobackupex continuing with FTWRL even when xtrabackup               failed
The problem was that when the log copying thread failed, it did not
terminate the xtrabackup process immediately. Instead only the log
copying thread was terminated. The main xtrabackup thread would later
check if the 'log_copying_succeed' flag is FALSE, and terminated the
process, but that happened later, after xtrabackup was done with copying
InnoDB files and signaled innobackupex to lock tables and proceed with
copying non-InnoDB files.

The fix is to terminate the entire xtrabackup process immediately on log
copying failure. There is no reasons to delay checking its status
and fail much later.