Implementation of https://blueprints.launchpad.net/percona-xtrabackup/+spec/parallel-table-rebuild
Add a new option to innobackupex/xtrabackup called --rebuild-threads
with the default value 1. The option specifies the number of threads
started by XtraBackup when rebuilding secondary indexes on --apply-log
--rebuild-indexes. Each thread rebuilds indexes for a single .ibd
tablespace at a time.

UNIV_LIST_DEBUG had to be removed from debug flags for non-5.6 build
configurations, as the UNIV_LIST_DEBUG implementation of
UT_LIST_REMOVE_CLEAR in 5.5- is not C++ compatible.