Bug #1177206: InnoDB: Assertion failure in thread <nr> in file               log0recv.c line 1105 InnoDB: Failing assertion: !page               || (ibool)!!page_is_comp(page) ==               dict_table_is_comp(index->table)
Bug #1187071: xb 2.1 against ps 5.6 gives Error: Tablespace is not
              sensible during backup"

The problems are related, so fixing in a single revision.

The problem in bug #1187071 was a difference in behavior between InnoDB
5.5- and 5.6 codebases in cases when a newly created tablespace has
uninitialized first page at the time when XtraBackup opens it while
creating a list of tablespaces to backup. 5.1/5.5 InnoDB code would just
complain about tablespace ID being non-sensible and omit it from the
backup. But in 5.6 it fails with an error in such a case.

A related problem in bug #1177206 was that
xb_data_files_init() (i.e. loading tablespaces into fil_system) was
called before log copying was started. This could lead to a race between
the datafiles state in the resulting backup and xtrabackup_logfile: a
tablespace created at a sensitive time would be missing in both the
backup itself and as the corresponding MLOG_FILE_CREATE* log record in
xtrabackup_logfile, so it would not be created on --apply-log
either. This problem has existed since the very first XtraBackup
version.

Fixed by:

1. Ignoring tablespaces with uninitialized pages on xtrabackup startup
for InnoDB 5.6 codebase to match the behavior of InnoDB 5.5-.
2. Moving the call to xb_data_files_init() in xtrabackup_backup_func()
to a later stage so that it is only called after the log copying is
started. This involved a minor code refactoring, since
xb_data_files_init() previously did 2 things: initializing the 'fil'
subsystem (which is also required by the log copying code) and actually
populating fil_system. We now have 2 separate functions:
xb_fil_io_init() (called before log copying is started) and
xb_load_tablespaces() (called after log copying is started).